refactor, move right drawer components
