+++ frontend/src/components/Workspace/FlowChart/Buttons/Snakemake.tsx
@@ -1,21 +1,10 @@
-import { memo, useEffect } from "react"
-import { useSelector, useDispatch } from "react-redux"
+import { memo } from "react"
+import { useDispatch } from "react-redux"
 
 import RouteIcon from "@mui/icons-material/Route"
 import { IconButton, Tooltip } from "@mui/material"
 
-import { createParamFormItemComponent } from "components/common/ParamFormItemCreator"
 import { toggleSnakemake } from "store/slice/RightDrawer/RightDrawerSlice"
-import { ParamItemProps } from "store/slice/RightDrawer/RightDrawerType"
-import { getSnakemakeParams } from "store/slice/Snakemake/SnakemakeAction"
-import {
-  selectSnakemakeParam,
-  selectSnakemakeParamsKeyList,
-  selectSnakemakeParamsValue,
-} from "store/slice/Snakemake/SnakemakeSelectors"
-import { updateParam } from "store/slice/Snakemake/SnakemakeSlice"
-import { AppDispatch } from "store/store"
-import { arrayEqualityFn } from "utils/EqualityUtils"
 
 export const SnakemakeButton = memo(function SnakemakeButton() {
   const dispatch = useDispatch()
@@ -29,34 +18,4 @@ export const SnakemakeButton = memo(function SnakemakeButton() {
       </IconButton>
     </Tooltip>
   )
-})
-
-export const SnakemakeContents = memo(function SnakemakeContents() {
-  const dispatch = useDispatch<AppDispatch>()
-  const paramKeyList = useSelector(
-    selectSnakemakeParamsKeyList,
-    arrayEqualityFn,
-  )
-  useEffect(() => {
-    if (paramKeyList.length === 0) {
-      dispatch(getSnakemakeParams())
-    }
-  })
-  return (
-    <div className="SnakemakeParam" style={{ margin: 4 }}>
-      {paramKeyList.map((paramKey, i) => (
-        <ParamItem key={i} paramKey={paramKey} />
-      ))}
-    </div>
-  )
-})
-
-const ParamItem = memo(function ParamItem({ paramKey }: ParamItemProps) {
-  const Component = createParamFormItemComponent({
-    paramSelector: selectSnakemakeParam,
-    paramValueSelector: selectSnakemakeParamsValue,
-    paramUpdateActionCreator: (path, newValue) =>
-      updateParam({ path, newValue }),
-  })
-  return <Component paramKey={paramKey} />
 })
